refactor(ledger): extract signature tests into a separate file
Move the inline `signature_tests` module from `mod.rs` into its own file to reduce the size of the main ledger module and improve test organization. The test logic is unchanged, only relocated via a `#[path]` attribute.

Auto-committed-on: dragonfly
